services/processor/requirements.txt
kafka-python==2.0.2
clickhouse-connect==0.7.0
